+++ Sources/CollectionViewModelCell.swift
@@ -18,6 +18,7 @@ public enum CollectionCellType {
     
 open class CollectionCellBasic: CollectionViewModelCell {
     public weak var delegate: ViewModelCellDelegate?
+    open var userInfo = [AnyHashable : Any]()
 
     public override init(frame: CGRect) {
         super.init(frame: frame)
@@ -41,7 +42,24 @@ open class CollectionCellBasic: CollectionViewModelCell {
     open func reset() {}
     open func update(with item: Item) {}
     
-    @objc public func performCallback(_ sender: Any) {
+    @objc open func performCallback(_ sender: Any) {
         delegate?.action(for: self, sender: sender)
+    }
+}
+
+open class CollectionCellStack: CollectionCellBasic {
+    public let stack = UIStackView()
+
+    open override func configure() {
+        super.configure()
+
+        contentView.addSubview(stack)
+        stack.translatesAutoresizingMaskIntoConstraints = false
+
+        let margins = contentView.layoutMarginsGuide
+        stack.leadingAnchor.constraint(equalTo: margins.leadingAnchor).isActive = true
+        stack.trailingAnchor.constraint(equalTo: margins.trailingAnchor).isActive = true
+        stack.topAnchor.constraint(equalTo: margins.topAnchor).isActive = true
+        stack.bottomAnchor.constraint(equalTo: margins.bottomAnchor).isActive = true
     }
 }